- trocar gráficos da fase #5 do castelo - trocar, onde possível, int por Uint8/Sint8 no código

+++ character/character.h
@@ -38,7 +38,7 @@ struct object_colision;
  *
  */
 struct st_spriteFrame {
-    int delay; // time in milisseconds this frame will be shown /**< TODO */
+    uint delay; // time in milisseconds this frame will be shown /**< TODO */
     graphicsLib_gSurface frameSurface; /**< TODO */
 	st_spriteFrame() {
 		frameSurface.gSurface = NULL;
@@ -663,7 +663,7 @@ protected:
 
     virtual bool can_air_dash();
 
-    short get_projectile_max_shots();
+    Uint8 get_projectile_max_shots();
 
 
 // members
